Don't be so strict about content-type
Pilot testing detected that CDS didn't process if Content-Type has a charset declared (e.g., "application/json charset=UTF-8").  This fixes that.

This also removes the requirement for stuff with FHIR data to use content-type `application/json+fhir` because that's technically incorrect, in that the body of the message is not actually FHIR.  Only the elements in the bodies `data` property are.  Now just detect  FHIR by inspecting the data.