Support sub-folders in localRepository and make "reset" explicit
Support loading libraries in sub-folders of localRepository.  If the same library and version is encountered, but has different content, log an error.

Also don't reset the localRepo every time load is called.  Instead `load` should be additive.  Expose a `reset` function to allow the localRepo memory to be explicitly reset.

And for consistency w/ PatientSource API code, rename "clear" to "reset" in localHooks code.