Card responses should have JSON values
When substituting in the CQL values, the result was always a string, despite the return type.  Now if the card config has a bare expression (e.g., "${myExp}"), replace it w/ the proper JSON value so that type is preserved.  Booleans stay booleans, arrays stay arrays, objects stay objects, etc.