Merged in update-deps-nov-22 (pull request #22)
Update Node, NPM, Docker, Dependencies, CQL Libraries, and Example Patients

* Upgrade Node, NPM, Docker, Dependencies

- Update Node to 16
- Switch Yarn to NPM 8
- Update Docker to Node 16
- Update dependency libraries to latest versions

* Update packaged CQL libraries and patients

Update the CQL libraries to their latest versions and remove the ASCVD library since we no longer maintain it.

Update the patients to R4 for use in the examples.

Update the client to use the updated libraries and patients.

Update the README to reflect all of the above.

* Update dependencies to resolve vulnerabilities


Approved-by: Julia Afeltra